Separated out the datatypes for readability
