+++ datatypes/config.go
@@ -1,0 +1,7 @@
+package datatypes
+
+// Config contains the config necessary for the app to run
+type Config struct {
+	Env    string
+	Server string
+}
